Resolve application overlap issues
Resolve issues caused by overlapping calls
or configuration in applications by including
an argument to invoked main methods indicating
that the application was started by the multi-
loader.

The invoked application is then responsible
for disabling the necessary features to avoid
conflict with other applications, for
instance, an application's auto restart
functionality or event handler creation.

Event handler creation is managed by the
multi-loader application now as it is unsafe
for an invoked application to create an
event handler itself. If duplicate event
handlers were created, infinite error logs
from the event handler warned of a duplicate
until the JVM hard crashed.